requirements.txt
flask==3.1.2
psycopg2-binary==2.9.10
mcp==1.13.1
python-dotenv==1.1.1
langchain==0.3.7
langchain-ollama==0.2.2